Add files via upload
Here's an alternative approach to achieve the same result using Python. This solution will automate the process of setting up the Docker containers to host the Minecraft servers.
In this approach, the Dockerfile is created and modified programmatically to use the specified Minecraft version and memory limit. The subprocess module is used to run the necessary Docker commands.